requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiokafka==0.8.0
aioredis==2.0.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
base58==2.1.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jsonalias==0.1.1
kafka-python==2.0.2
kombu==5.5.2
multidict==6.4.3
packaging==24.2
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions>=4.8.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
vine==5.1.0
wcwidth==0.2.13
websockets==15.0
yarl==1.19.0
pandas>=1.5.0